requirements.txt
astroid==2.0.4
click==6.7
colorama==0.3.9
docutils==0.14
first==2.0.1
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Jsonpify==1.5.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.15.1
pandas==0.20.3
pathlib==1.0.1
pip-tools==1.9.0
pylint==2.1.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.13
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.1.12
statistics==1.0.3.5
typed-ast==1.1.0
Werkzeug==0.12.2
wrapt==1.10.11
